examples/requirements.txt
transformers==4.56.*
huggingface-hub
diffusers==0.32.0
sentencepiece
pillow
pytest-forked
tiktoken
blobfile
torchvision
omegaconf
